enhance: upgrade Vortex to 0.75.0 using a forked repository (#588)
We've been carrying Vortex 0.56 through a locally cloned and patched
workspace so Milvus can use inline flat-layout sub-segment reads. That
setup keeps the bridge tied to an older Vortex API and makes every build
run a custom patching step. This change moves the dependency to the
0.75-based feature branch in the Vortex 0.75-based feature branch, where
the sub-segment read changes and the public StatsAccumulator entry point
are maintained together. Point the Rust bridge directly at
feature/sub-segment-read-075, refresh the full Vortex crate set, and
remove the CMake clone-and-patch flow, generated checkout directory,
shell script, and checked-in patch files. Keep both Arrow generations in
the bridge and move only the Vortex conversion path to Arrow 58,
avoiding an unrelated bridge-wide Arrow migration.

Adapt the writer, scanner, custom V2 layout, and statistics pruning
paths to Vortex 0.75's session-based APIs. Thread VortexSession and
ExecutionCtx through layout readers and writers, switch recursive struct
writing to TableStrategy, keep the same BtrBlocks compression choices,
and use the new accumulator finalization and expression execution APIs
without changing the row-group split policy, per-column zone maps,
boundary metadata, or conservative pruning behavior. For object-store
reads, replace the old push-driven ReadSource implementation with
VortexReadAt while preserving the existing C++ filesystem handles,
asynchronous and blocking FFI paths, coalescing settings, concurrency
limit, tracing, and error mapping. This lets Vortex own request
scheduling and buffer handling while Milvus keeps the same read
semantics and sub-segment planning behavior. The vortex V1 and V2 writer
and reader contracts remain the same, including projections, take scans,
footer caching, and empty writes. Vortex 0.75 can emit schema and layout
segments even for a zero-row file, so footer validation now uses the
parsed footer byte range instead of assuming that every byte after the
four- byte header belongs to the footer.

---------

+++ cpp/src/format/bridge/rust/src/lib.rs
@@ -328,7 +328,7 @@ pub mod vortex_ffi {
         fn with_row_range(self: &mut VortexScanBuilder, row_range_start: u64, row_range_end: u64);
         fn with_row_ranges(self: &mut VortexScanBuilder, starts: &[u64], ends: &[u64]);
         fn with_include_by_index(self: &mut VortexScanBuilder, include_by_index: &[u64]);
-        fn with_limit(self: &mut VortexScanBuilder, limit: usize);
+        fn with_limit(self: &mut VortexScanBuilder, limit: u64);
         unsafe fn with_output_schema(
             self: &mut VortexScanBuilder,
             output_schema: *mut u8,
